backend/requirements.txt
Flask
flask-cors
requests
supabase
python-dotenv